LPS-106315 Inline to where it is used
