+++ modules/dxp/apps/oauth/oauth-service/src/main/java/com/liferay/oauth/internal/V10aOAuth.java
@@ -15,7 +15,6 @@
 package com.liferay.oauth.internal;
 
 import com.liferay.oauth.configuration.OAuthConfigurationValues;
-import com.liferay.oauth.internal.listener.V10aOAuthDebugCacheListener;
 import com.liferay.oauth.model.OAuthApplication;
 import com.liferay.oauth.model.OAuthUser;
 import com.liferay.oauth.service.OAuthApplicationLocalService;
@@ -31,6 +30,8 @@ import com.liferay.oauth.util.OAuthConsumer;
 import com.liferay.oauth.util.OAuthMessage;
 import com.liferay.oauth.util.OAuthValidator;
 import com.liferay.portal.kernel.cache.PortalCache;
+import com.liferay.portal.kernel.cache.PortalCacheException;
+import com.liferay.portal.kernel.cache.PortalCacheListener;
 import com.liferay.portal.kernel.cache.SingleVMPool;
 import com.liferay.portal.kernel.cluster.ClusterExecutor;
 import com.liferay.portal.kernel.cluster.ClusterRequest;
@@ -50,6 +51,7 @@ import com.liferay.portal.kernel.util.MethodHandler;
 import com.liferay.portal.kernel.util.MethodKey;
 import com.liferay.portal.kernel.util.Portal;
 import com.liferay.portal.kernel.util.PwdGenerator;
+import com.liferay.portal.kernel.util.StringBundler;
 import com.liferay.portal.kernel.util.Validator;
 
 import java.io.IOException;
@@ -427,5 +429,88 @@ public class V10aOAuth implements IdentifiableOSGiService, OAuth {
 
 	@Reference
 	private SingleVMPool _singleVMPool;
+
+	private static class V10aOAuthDebugCacheListener<K extends Serializable, V>
+		implements PortalCacheListener<K, V> {
+
+		@Override
+		public void dispose() {
+		}
+
+		@Override
+		public void notifyEntryEvicted(
+				PortalCache<K, V> portalCache, K key, V value, int timeToLive)
+			throws PortalCacheException {
+
+			logDebug("Entry evicted", key, value);
+		}
+
+		@Override
+		public void notifyEntryExpired(
+				PortalCache<K, V> portalCache, K key, V value, int timeToLive)
+			throws PortalCacheException {
+
+			logDebug("Entry expired", key, value);
+		}
+
+		@Override
+		public void notifyEntryPut(
+				PortalCache<K, V> portalCache, K key, V value, int timeToLive)
+			throws PortalCacheException {
+
+			logDebug("Entry put", key, value);
+		}
+
+		@Override
+		public void notifyEntryRemoved(
+				PortalCache<K, V> portalCache, K key, V value, int timeToLive)
+			throws PortalCacheException {
+
+			logDebug("Entry removed", key, value);
+		}
+
+		@Override
+		public void notifyEntryUpdated(
+				PortalCache<K, V> portalCache, K key, V value, int timeToLive)
+			throws PortalCacheException {
+
+			logDebug("Entry updated", key, value);
+		}
+
+		@Override
+		public void notifyRemoveAll(PortalCache<K, V> portalCache)
+			throws PortalCacheException {
+
+			if (_log.isDebugEnabled()) {
+				_log.debug("Remove all " + portalCache.getPortalCacheName());
+			}
+		}
+
+		protected void logDebug(String method, K key, V value) {
+			if (!_log.isDebugEnabled()) {
+				return;
+			}
+
+			if (!(value instanceof OAuthAccessor)) {
+				return;
+			}
+
+			OAuthAccessor oAuthAccessor = (OAuthAccessor)value;
+
+			StringBundler sb = new StringBundler(7);
+
+			sb.append(method);
+			sb.append("  ");
+			sb.append(key);
+			sb.append(":");
+			sb.append(oAuthAccessor.getRequestToken());
+			sb.append(":");
+			sb.append(
+				oAuthAccessor.getProperty(OAuthAccessorConstants.AUTHORIZED));
+
+			_log.debug(sb.toString());
+		}
+
+	}
 
 }
